✨ refactor(modal): remove file structure
Add functions for creating and managing modals, including `createModal`, `createRawModal`, and their respective bridge variants. Introduce a modal stack for tracking active modals and operations for updating, closing, and destroying modals. Implement a `ModalHost` and `ModalPortal` for rendering modals in a React application, enhancing the modal system's flexibility and usability.

- Create `createModal` and `createRawModal` functions for modal instantiation
- Implement `ModalHost` and `ModalPortal` components for rendering
- Add utility functions for modal stack management and operations
- Define types for modal properties and stack items
- Ensure modals can be updated and closed programmatically
